Rename some files to match their class name
